Remove bool operators and implicit cast to bool
These made more sense with the Maybe<T> class that I originally created.  Maybe<T> included, along with the ability to contain an exception, the ability to be empty.  The bool operators and conversion allowed one to do this:

Maybe<int> value = SomeMethod();

if (value)
{
    // Use value
}

This made more sense with checking for empty, than for the absence of an exception.

It does mean that one has to call HasValue, but I think it makes sense to  be more explicit when it comes to exceptions, and less 'magical'.

Also it would have been confusing if creating an Exceptionable<bool>.